Refactor logging utilities into separate modules
